refactor: harden canvas routes
chore: move canvas types to shared lib
