+++ apps/api/src/main.ts
@@ -349,14 +349,7 @@ if (ENV.NODE_ENV === 'production' && !ENV.CI && cluster.isPrimary) {
   app.use('/web-extension', setCacheControlForApiRoutes, webExtensionRoutes);
   app.use('/desktop-app', setCacheControlForApiRoutes, desktopAppRoutes);
   app.use('/openapi', setCacheControlForApiRoutes, openApiRoutes);
-  app.use(
-    '/canvas',
-    (_: express.Request, res: express.Response, next: express.NextFunction) => {
-      res.setHeader('Content-Security-Policy', "frame-ancestors 'self' *.force.com *.salesforce.com *.salesforce-setup.com;");
-      next();
-    },
-    canvasRoutes,
-  );
+  app.use('/canvas', canvasRoutes);
 
   if (ENV.ENVIRONMENT !== 'production' || ENV.CI) {
     app.use('/test', testRoutes);
